// constants/types.ts
export interface User {
    id: string;
    name: string;
    email: string;
}

export interface ChatMessage {
    id: string;
    text: string;
    isUser: boolean;
    timestamp: Date;
    UserName: string;
}

export interface ChatRoom {
    id: string;
    name: string;
    description: string;
    icon: string;

}